fix(NPE): Prevent crashes when clicking HAC toggle in search result
The crahes happen when the toggle is shown in the Settings search result
without proper initialization on the controller. This toggle will be
removed in the near future as Settings team is removing slice related
feature. At the mean time, we'll add a simple null check to prevent
crashes happen.

Flag: EXEMPT BUGFIX
Bug: 457759139
Test: manually test
Change-Id: I1492c20d3fabc21dfbfa2d65e5da725009444150