RTL ColorModePreference photo swiping improvement
Added a workaround for right-to-left languages swiping options among the color mode preferences. This is a temporary solution to the current usage of ViewPager which does not have automatic mirroring.

The photos are in reverse order compared to the default LTR languages and start on the right-most page.

Bug: 435135298
Flag: EXEMPT BUGFIX
Test: http://screencast/cast/NjI0Mjc0NDA2NTMyNzEwNHwwOGYyMzA4ZS04Yw
NO_IFTTT=no changes to the logic in related files

Change-Id: I6858ed82e75e7f00ad80771ce38269e54f5d1baa