Only show desktop mode option and its settings when desktop is supported
In SystemUI: when connecting a display, only show the desktop/extend
option in the connection preference dialog when desktop mode is
supported.

In DisplayManager:
1 - On startup, override the mirror/extend setting to be mirror if
desktop mode is not supported.
2 - When returning the connection preference, check if desktop mode is
supported, and override it if it was desktop/extend, and desktop mode is
not supported.

In Settings: only show mirror/extend related settings if desktop mode is
supported.

Bug: 450957248
Flag: EXEMPT Bugfix
Test: TBD
Change-Id: I11107198dcb86f213f1fe723dd154e763f296e70