Check for ADMIN in mobile network restriction check
The original check for the GUEST user is not restrictive enough, as
all FULL users who are not ADMIN would be allowed and not restricted,
and this goes against the general guidance of who is allowed to manage
Telephony.

Check for whether the user is ADMIN instead, and disallow right away
if it isn't.

Bug: 464248646
Test: atest MobileNetworkUtilsTest
Test: atest CreateShortcutPreferenceControllerTest
Test: atest DataUsageSummaryTest
Test: atest SimDialogActivityTest
Flag: EXEMPT BUGFIX

Change-Id: Id084d1784b4f315ff042b4800af41017e66bbeae