Migrate the multiuser WiFi feature to use new aconfig flag
The multiuser Wi-Fi setting feature is a pillar agnostic feature which
should use a generic platform namespace.

To migrate the old flag in the desktop_connectivity namespace to
android_settings namespace, 3 steps are needed. This change is the 2nd
step in flag migration,

1. Create a new flag in the android_settings namespace.
2. Move the feature to use the new flag.
3. When the new flag reaches same stage in gantry, remove the old flag
   to ensure smooth transition.

NO_IFTTT=The non-Admin restriction is not implemented in
WifiHotspotScreen.kt so no need to modify
Bug: 402214828
Flag: com.android.settings.flags.enable_wifi_multiuser
Test: atest SettingsRoboTests:com.android.settings.wifi
Test: on device test, check the feature is enabled

Change-Id: I29ca99c97e4426a09f867da71389e016345107f8