Add separate Show Secrets settings for touch and physical inputs
This change implements the UI support for the new split "Show Secrets"
settings, allowing users to independently control whether passwords and
other secrets are briefly displayed when typing on a touchscreen versus
on a physical keyboard.

If the flag and compat change are enabled the exisiting
`ShowPasswordPreference` setting will be hidden in favor of the two new
ones.

Bug: 453646492
Test: atest SettingsRoboTests:com.android.settings.security.ShowPasswordPreferenceControllerTest
Test: atest SettingsRoboTests:com.android.settings.security.ShowSecretsTouchPreferenceControllerTest
Test: atest SettingsRoboTests:com.android.settings.security.ShowSecretsPhysicalPreferenceControllerTest
Flag: com.android.text.flags.split_show_passwords_to_touch_and_physical
Change-Id: Ia8f5125fcce49fb65617d92190f2c6cead8c8532