fix(mag always on): the always on toggle should be default on
In AlwaysOnSwitchPreference, we didn't make the toggle default on,
causing an inconsistency that in a fresh build the toggle is off
but the accessibility platform have the functionality enabled.

Fix the inconsitency, also we remove the feature flag so now the
preference only check the config for the visibility.

Bug: 472478320
Flag: EXEMPT BUGFIX
Test: manually check
      atest AlwaysOnSwitchPreferenceTest
      atest AlwaysOnPreferenceControllerTest
Change-Id: I0113fa7a8e9fd8dff3e9c09355170a93cbce4b7e